Add ControlLabel and update ItemCountLabel
